move top level commands to /cmd/swarmctl/root/
