+++ cmd/swarmctl/root/common.go
@@ -1,4 +1,4 @@
-package main
+package root
 
 import (
 	"os"
