+++ cmd/swarmctl/root/deploy.go
@@ -1,4 +1,4 @@
-package main
+package root
 
 import (
 	"fmt"
